per DB's patch, convert placewewant to a size_t; also add a few miscellaneous fixes of mine
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1870 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/global.c
@@ -91,7 +91,7 @@ int resetstatuspos;		/* Hack for resetting the status bar
 char *answer = NULL;		/* Answer str to many questions */
 int totlines = 0;		/* Total number of lines in the file */
 long totsize = 0;		/* Total number of bytes in the file */
-int placewewant = 0;		/* The column we'd like the cursor
+size_t placewewant = 0;		/* The column we'd like the cursor
 				   to jump to when we go to the
 				   next or previous line */
